refactor: split simulation runner modules
